% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipit.R
\name{gzipit}
\alias{gzipit}
\title{gzipit}
\usage{
gzipit(filepath)
}
\arguments{
\item{filepath}{the path to the file}
}
\description{
zip file with gzip
}
\examples{
foo <- tempfile(fileext = ".txt")
cat(file = foo, "Hello world!")
gzipit(foo)
}
