% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipit.R
\name{zipit}
\alias{zipit}
\title{zipit}
\usage{
zipit(filepath)
}
\arguments{
\item{filepath}{the path to the file}
}
\description{
zip file with zip
}
\examples{
foo <- tempfile(fileext = ".txt")
cat(file = foo, "Hello world!")
zipit(foo)
}
