% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziptest-package.R
\docType{package}
\name{ziptest-package}
\alias{ziptest}
\alias{ziptest-package}
\title{ziptest: tests zip problems on WIN DEVEL}
\description{
tests zip problems on WIN DEVEL. not a real package.
}
\author{
\strong{Maintainer}: Stefan Mayer \email{stefan@mayer-de.com}

}
\keyword{internal}
